Initial port of unittest test_junos to pytest
